assistant/requirements.txt
python-dotenv~=1.0.0
requests~=2.31.0
openai~=0.27.8
fastapi==0.100.0
uvicorn==0.23.1